+++ src/help.rs
@@ -18,13 +18,15 @@ USAGE:
 FLAGS:
     --help                       # Print this message and exit
     --version                    # Print version string and exit
+    --quiet                      # Reduce amount of stderr logs
     --watch                      # Rerun tests on file changes
     --compiler /path/to/compiler # Precise the compiler to use (defaults to just elm)
     --seed integer               # Run with initial fuzzer seed (defaults to random)
     --fuzz integer               # Precise number of iterations of fuzz tests (defaults to 100)
     --workers integer            # Precise number of worker threads (defaults to number of logic cores)
     --filter "substring"         # Keep only the tests whose descriptions contain the given string
-    --report console|json|junit  # Print results to stdout in given the format (defaults to console)
+    --report console|json|junit|exercism
+                                 # Print results to stdout in given the format (defaults to console)
     --connectivity progressive|offline|online-newest|online-oldest
                                  # Connectivity mode (defaults to progessive)
                                  #    offline: elm-test-rs only use installed packages to solve dependencies